requirements.txt
# Django and core dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0

# Celery and Redis
celery>=5.3,<6.0
redis>=5.0,<6.0

# AWS S3
boto3>=1.34,<2.0
botocore>=1.34,<2.0

# Data processing
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0
xlrd>=2.0,<3.0

# File handling
python-magic>=0.4,<1.0

# HTTP and CORS
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Filtering and utilities
django-filter>=23.5,<24.0

# Development tools (optional)
ipython>=8.0,<9.0
django-extensions>=3.2,<4.0

# Testing (optional)
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0

